requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22